Removed factory.py
Finally managed to track down what was wrong with reusing Estimators,
which was what made factories necessary in the first place. (Namely, it
was the same Layer objects being used each time model_fn (defined in
Network.compile) was called.) The fix for this is to doubly encapsulate
the layer objects before adding them to the Network. (As one layer of
encapsulating gets removed when defining the layer we wish to add.)
Which is all a bit silly really. (Searches online suggest that the other
potential fix (setting reuse=True on a layer) is discouraged.)

Also made some other changes:

-Renamed Sequential to Network, as it doesn't operate in quite the same
manner as keras.Sequential. It should be viewed as defining the layout
of the graph, but without populating it with weights and such yet.

-Added train.py, which allows for training a model adaptively, i.e.
automatically decreasing the learning rate when necessary.

-The processors for Network should now also be lambda-wrapped, so that
they can be recreated each model_fn call. (Similar to the Layers).

-Added Network.define_dnn, for ease of creating simple networks in a
manner similar to the old factories.

-Moved processor inverse transforms to inside model_fn, where they
always really should have been. This also means that the expected labels
are no longer transformed by processors, as they don't need to be.

-Networks may now have default compile kwargs registered.